% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.R
\name{cbind.metamet}
\alias{cbind.metamet}
\title{Combine metamet Objects by Columns}
\usage{
\method{cbind}{metamet}(..., deparse.level = 1)
}
\arguments{
\item{...}{metamet objects to combine}

\item{deparse.level}{Integer determining the naming of combined data tables
(unused for metamet objects, included for S3 method compatibility)}
}
\value{
A new metamet object containing:
  \item{dt}{Combined data table with columns from all input objects,
    merged by TIMESTAMP}
  \item{dt_meta}{Unique rows from metadata tables of all input objects}
  \item{dt_site}{Unique rows from site information tables of all input objects}
}
\description{
Combines multiple metamet objects by column-binding their data tables,
while merging metadata tables by variable name.
}
\examples{
\dontrun{
  mm_combined <- cbind(mm1, mm2, mm3)
}

}
