% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_data.R
\name{detect_gaps}
\alias{detect_gaps}
\title{detect_gaps}
\usage{
detect_gaps(dt_in, expected_interval_mins = NULL, time_name = NULL)
}
\arguments{
\item{time_name}{Column name for POSIX date/time variable in df}

\item{df}{A data frame}

\item{by}{Time interval of series, Default: '30 min'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Detects any gaps in a data frame representing a time series
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 gaps <- detect_gaps(l_logr$df)
 }
}
}
