% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamet.R
\name{dummy}
\alias{dummy}
\title{dummy}
\usage{
dummy(x = 1:10)
}
\arguments{
\item{x_in}{url to the template repository}
}
\description{
A dummy function for testing
}
\details{
none
}
\examples{
\dontrun{
dummy(1:5)
}
}
