% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggiraph_plot}
\alias{ggiraph_plot}
\title{ggiraph_plot}
\usage{
ggiraph_plot(input_variable)
}
\arguments{
\item{input_variable}{The name of the variable within the query data frame to plot.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Creates an interactive girafe plot, whereby the user can select
  points with dubious quality and impute new values.
}
\details{
Custom plotting function for each variable

DETAILS
}
