% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamet-package.r
\docType{package}
\name{metamet-package}
\alias{metamet-package}
\title{metamet: Meta Data for Met Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package defines a data structure for met data and its meta data, and functions for manipulation and quality control.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nerc-ceh.github.io/metamet/}
}

}
\author{
\strong{Maintainer}: Peter Levy \email{plevy@ceh.ac.uk} (\href{https://orcid.org/0000-0002-8505-1901}{ORCID}) [copyright holder]

}
\keyword{internal}
