% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_data.R
\name{pad_data}
\alias{pad_data}
\title{pad_data}
\usage{
pad_data(dt, by = NULL, time_name = NULL, v_dates = NULL)
}
\arguments{
\item{by}{Time interval of series, Default: '30 min'}

\item{time_name}{Column name for POSIX date/time variable in df, Default: 'DATECT'}

\item{v_dates}{A vector of POSIX date/times, potentially from another df, to match with it.}

\item{df}{A data frame}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Adds in any gaps in a data frame representing a time series
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 df <- pad_data(df)
 }
}
}
