% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Launches the shiny app for the metamet package}
\usage{
run_shiny()
}
\value{
shiny application object
}
\description{
This function provides a web browser interface.
}
\keyword{app}
\keyword{shiny}
