% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_date.R
\name{subset_by_date}
\alias{subset_by_date}
\title{Subset a `metamet` object by date}
\usage{
subset_by_date(
  mm,
  start_date = "2025-01-01 00:30:00",
  end_date = "2026-01-01 00:00:00"
)
}
\arguments{
\item{mm}{A `metamet` object containing observation data and metadata.}

\item{start_date}{Earliest date to include in subset.}

\item{end_date}{Last date to include in subset.}
}
\value{
The `metamet` object `mm`, restricted to between start and end dates.
}
\description{
Subsets a `metamet` object to a date range between a start and end date.
This is applied to the observation data (`dt`) and the QC (`dt_qc`) and
reference data (`dt_ref`) if they are present.
}
\details{
- Limits all time-referenced data to between the start and end dates.
- Dates are inclusive.
- Warns if date variable not present.
}
\examples{
\dontrun{
mm <- subset_by_date(
  mm,
  start_date = "2025-06-01 00:30:00",
  end_date   = "2025-06-02 00:00:00"
)
}
}
\seealso{
- \code{\link{rename_era5}} for variable renaming
- \code{\link{time_average_dt}} for time-averaging implementation
- \code{\link{metamet}} for object construction
}
