% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{impute}
\usage{
impute(
  y,
  mm,
  method = "era5",
  qc_tokeep = 0,
  selection = TRUE,
  date_field = "DATECT",
  k = 40,
  fit = TRUE,
  n_min = 10,
  x = NULL,
  lat = 55.792,
  lon = -3.243,
  plot_graph = TRUE
)
}
\arguments{
\item{y}{Response variable with missing values to be replaced
(variable name as a "quoted string")}

\item{mm}{metamet object containing data and qc codes.}

\item{fit}{Whether to fit a linear model or directly replace missing y with
x values, Default: TRUE}

\item{x}{Covariate to be used (name of a variable in the same data frame as
a "quoted string")}

\item{qc}{qc code to denote values imputed by this function, Default: 3}
}
\value{
List of two data frames containing data and qc codes with imputed values.
}
\description{
Impute missing values using various methods
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()) {
 #EXAMPLE1
mm <- list(df = df, df_qc = df_qc)
mm <- impute(y = "SW_IN", x = "PPFD_IN",  mm)
 }
}
}
