% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{rbind_metamet}
\alias{rbind_metamet}
\title{Combine lists of metamet data tables into a single metamet object}
\usage{
rbind_metamet(mm, l_dt, l_dt_meta, l_dt_site)
}
\arguments{
\item{mm}{A `metamet` object to receive the combined tables.}

\item{l_dt}{A list of data tables to row-bind into `mm$dt`.}

\item{l_dt_meta}{A list of metadata tables to row-bind into `mm$dt_meta`.}

\item{l_dt_site}{A list of site tables to row-bind into `mm$dt_site`.}
}
\value{
The modified `metamet` object.
}
\description{
Row-binds lists of data tables into an existing `metamet` object. Typically
used to merge data from multiple sites or time periods.
}
\examples{
\dontrun{
mm <- rbind_metamet(mm, list(dt1, dt2), list(meta1, meta2), list(site1, site2))
}

}
